Fix bug with speedratio and floats

+++ src/abb_robot_client/rws.py
@@ -793,7 +793,7 @@ class RWS:
             raise Exception("Invalid speedratio type")
         return float(state["speedratio"])
     
-    def set_speedratio(self, speedratio: float):
+    def set_speedratio(self, speedratio: int):
         """
         Set the current speed ratio
